Use memcached for updated_at and last_signed_at too
We see locks on timestamping the petition after each signature validation on high traffic petitions.  To avoid this we stop writing this to the db, but we still need to change the updated_at because it's used in the cache_key.  Instead of trying to work out which signature validations should trigger a write we follow the pattern established by signature_count and put the updated_at and last_signed_at into memcached.  We change the CachedSignatureCountResetJob to be CachedPetitionValuesResetJob and write all the cached values back into the db periodically.

Note that we use change(nsec: 0) in our specs when dealing with time to avoid database precision issues.